+++ geolocator/CHANGELOG.md
@@ -1,3 +1,20 @@
+
+## 7.6.1
+
+- Refactored the example app and rollbacked the apple deployment target from 9.0 to 8.0.
+
+## 7.6.0
+
+- Make sure the `getCurrentPosition` method returns the current position and not a cached location which might be wrong (see issue [#629](https://github.com/Baseflow/flutter-geolocator/issues/629)). Note that this means that in some cases fetching the current position might take several seconds. If you want to have a fast initial result, call the `getLastKnownPosition` first and update the position with the result from `getCurrentPosition`.
+
+## 7.5.0
+
+- Added support for macOS Desktop.
+
+## 7.4.1
+
+- Updated README.md to include information about using the Geolocator web version in release mode.
+
 ## 7.4.0
 
 - Moved to fully federated architecture (moved Android and iOS implementations into separate packages, geolocator_android and geolocator_apple respectivily).