+++ geolocator_apple/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.2.0
+
+- Make sure the `getCurrentPosition` method returns the current position and not a cached location which might be wrong (see issue [#629](https://github.com/Baseflow/flutter-geolocator/issues/629)).
+
+## 1.1.0
+
+- Added support for macOS Desktop.
+
 ## 1.0.0
 
 - Initial open source release.